feat: Add complete S41D implementation for C1 as py file chore: Add finalised basic mvp S41D implementation as ipynb chore: Update gitignore
The complete implementation will try to download the requisite data from
eegdash if the data is not already present and downloaded in a data/
folder at the project root. Logs will be placed in logs/ but are in the
gitignore so they won't sync. Same with the eeg-pretraining private repo
submodule, and archive of old files